story: modeler extension (#1279)
* Updated bpmn modeler to latest version

* Added custom renderer

* Updated other BPMN viewers

* WIP

* WIP

* WIP

* upgrade and lock deps

* work on editor

* work on editor

* version switch, progress

* read only tag

* only switch version when different from current

* disable overflow

* system tags

* use navigated viewer instead of viewer

* panel, fix styling

* no changes pending on version switch

* small styling improvement

* wip

* refactor to service

* work on accessing service in provider

* wip

* work on edit event

* working open edit modal

* update bpmn view

* unlink click

* open modal on create event

* respond to create event, implement for form flow

* wip

* cleanup

* improvements

* add export func

* work on create

* upload and create carbon list

* disable dark mode

* clean up

* try to fix migration

* fix lint

* fix deps

* fix migration screen

* include properties panel for styling again

* pr comments

* carbonize upload modal

* comment

---------

+++ projects/valtimo/process-link/src/lib/components/process-link-modal/process-link-modal.component.ts
@@ -24,6 +24,7 @@ import {
 } from '../../services';
 import {take} from 'rxjs/operators';
 import {ConfigService} from '@valtimo/config';
+import {ProcessLinkEditMode} from '../../models';
 
 @Component({
   selector: 'valtimo-process-link-modal',
@@ -81,6 +82,11 @@ export class ProcessLinkModalComponent {
     this.stateService.startSaving();
 
     this.stateService.selectedProcessLink$.pipe(take(1)).subscribe(selectedProcessLink => {
+      if (this.stateService.processLinkEditMode === ProcessLinkEditMode.EMIT_EVENTS) {
+        this.stateService.sendProcessLinkDeleteEvent({processLinkId: selectedProcessLink.id});
+        return;
+      }
+
       this.processLinkService.deleteProcessLink(selectedProcessLink.id).subscribe(
         () => {
           this.stateService.closeModal();
